chore/routing_model: extract out actions and utilities
- Extract out actions and utilities to their own files.
- Add copyright notices.
