+++ routing_model/src/lib.rs
@@ -1,11 +1,31 @@
+// Copyright 2019 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+// This is used two ways: inline tests, and integration tests (with mock).
+// There's no point configuring each item which is only used in one of these.
+
 #![allow(dead_code)]
 #![allow(unused_imports)]
+
+mod actions;
+mod utilities;
 
 use itertools::Itertools;
 use std::cell::RefCell;
 use std::collections::BTreeMap;
 use std::fmt::{self, Debug, Display, Formatter};
 use std::rc::Rc;
+
+use crate::actions::{Action, InnerAction};
+use crate::utilities::{
+    Age, Attributes, Candidate, ChangeElder, Event, GenesisPfxInfo, LocalEvent, Name, Node,
+    NodeChange, NodeState, ParsecVote, Proof, ProofRequest, ProofSource, Rpc, Section, SectionInfo,
+};
 
 #[macro_use]
 extern crate lazy_static;
@@ -719,635 +739,6 @@ impl JoiningRelocateCandidate {
 //////////////////
 /// Utilities
 //////////////////
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Eq, Ord)]
-struct Name(i32);
-
-#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Eq, Ord)]
-struct Age(i32);
-
-#[derive(Debug, Clone, Copy, Default, PartialEq)]
-struct Attributes {
-    age: i32,
-    name: i32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-struct Candidate(Attributes);
-
-#[derive(Debug, Clone, Copy, Default, PartialEq)]
-struct Node(Attributes);
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum NodeChange {
-    AddResourceProofing(Node),
-    Online(Node),
-    Relocating(Node),
-    Remove(Node),
-    Elder(Node, bool),
-}
-
-impl NodeChange {
-    fn node(&self) -> Node {
-        match &self {
-            NodeChange::AddResourceProofing(node)
-            | NodeChange::Online(node)
-            | NodeChange::Relocating(node)
-            | NodeChange::Remove(node)
-            | NodeChange::Elder(node, _) => *node,
-        }
-    }
-
-    fn relocating(&self) -> bool {
-        match &self {
-            NodeChange::Relocating(_) => true,
-            _ => false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq)]
-struct NodeState {
-    node: Node,
-    is_elder: bool,
-    is_relocating: bool,
-    need_relocate: bool,
-    is_resource_proofing: bool,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, PartialOrd, Ord, Eq)]
-struct Section(i32);
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, PartialOrd, Ord, Eq)]
-struct SectionInfo(Section, i32 /*contain full membership */);
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, PartialOrd, Ord, Eq)]
-struct GenesisPfxInfo(SectionInfo);
-
-#[derive(Debug, Clone, PartialEq)]
-struct ChangeElder {
-    changes: Vec<(Node, bool)>,
-    new_section: SectionInfo,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-struct ProofRequest {
-    value: i32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum Proof {
-    ValidPart,
-    ValidEnd,
-    Invalid,
-}
-
-impl Proof {
-    fn is_valid(&self) -> bool {
-        match self {
-            Proof::ValidPart | Proof::ValidEnd => true,
-            Proof::Invalid => false,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Default, Copy, Clone)]
-struct ProofSource(i32);
-
-impl ProofSource {
-    fn next_part(&mut self) -> Proof {
-        if self.0 > -1 {
-            self.0 -= 1;
-        }
-
-        self.resend()
-    }
-
-    fn resend(&self) -> Proof {
-        if self.0 > 0 {
-            Proof::ValidPart
-        } else if self.0 == 0 {
-            Proof::ValidEnd
-        } else {
-            Proof::Invalid
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum Event {
-    Rpc(Rpc),
-    ParsecConsensus(ParsecVote),
-    LocalEvent(LocalEvent),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum Rpc {
-    RefuseCandidate(Candidate),
-    RelocateResponse(Candidate, SectionInfo),
-    RelocatedInfo(Candidate, SectionInfo),
-
-    ExpectCandidate(Candidate),
-    ResendExpectCandidate(Section, Candidate),
-
-    ResourceProof {
-        candidate: Candidate,
-        source: Name,
-        proof: ProofRequest,
-    },
-    ResourceProofReceipt {
-        candidate: Candidate,
-        source: Name,
-    },
-    NodeApproval(Candidate, GenesisPfxInfo),
-
-    ResourceProofResponse {
-        candidate: Candidate,
-        destination: Name,
-        proof: Proof,
-    },
-    CandidateInfo {
-        candidate: Candidate,
-        destination: Name,
-        valid: bool,
-    },
-
-    ConnectionInfoRequest {
-        source: Name,
-        destination: Name,
-        connection_info: i32,
-    },
-    ConnectionInfoResponse {
-        source: Name,
-        destination: Name,
-        connection_info: i32,
-    },
-}
-
-impl Rpc {
-    fn to_event(&self) -> Event {
-        Event::Rpc(*self)
-    }
-
-    fn candidate(&self) -> Option<Candidate> {
-        match self {
-            Rpc::RefuseCandidate(candidate)
-            | Rpc::RelocateResponse(candidate, _)
-            | Rpc::RelocatedInfo(candidate, _)
-            | Rpc::ExpectCandidate(candidate)
-            | Rpc::ResendExpectCandidate(_, candidate)
-            | Rpc::ResourceProof { candidate, .. }
-            | Rpc::ResourceProofReceipt { candidate, .. }
-            | Rpc::NodeApproval(candidate, _)
-            | Rpc::ResourceProofResponse { candidate, .. }
-            | Rpc::CandidateInfo { candidate, .. } => Some(*candidate),
-
-            Rpc::ConnectionInfoRequest { .. } | Rpc::ConnectionInfoResponse { .. } => None,
-        }
-    }
-
-    fn destination(&self) -> Option<Name> {
-        match self {
-            Rpc::RefuseCandidate(_)
-            | Rpc::RelocateResponse(_, _)
-            | Rpc::RelocatedInfo(_, _)
-            | Rpc::ExpectCandidate(_)
-            | Rpc::ResendExpectCandidate(_, _)
-            | Rpc::NodeApproval(_, _) => None,
-
-            Rpc::ResourceProof { candidate, .. } | Rpc::ResourceProofReceipt { candidate, .. } => {
-                Some(Name(candidate.0.name))
-            }
-
-            Rpc::ResourceProofResponse { destination, .. }
-            | Rpc::CandidateInfo { destination, .. }
-            | Rpc::ConnectionInfoRequest { destination, .. }
-            | Rpc::ConnectionInfoResponse { destination, .. } => Some(*destination),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum ParsecVote {
-    ExpectCandidate(Candidate),
-    Online(Candidate),
-    PurgeCandidate(Candidate),
-    AddElderNode(Node),
-    RemoveElderNode(Node),
-    NewSectionInfo(SectionInfo),
-
-    RelocationTrigger,
-    RefuseCandidate(Candidate),
-    RelocateResponse(Candidate, SectionInfo),
-
-    CheckElder,
-}
-
-impl ParsecVote {
-    fn to_event(&self) -> Event {
-        Event::ParsecConsensus(*self)
-    }
-
-    fn candidate(&self) -> Option<Candidate> {
-        match self {
-            ParsecVote::ExpectCandidate(candidate)
-            | ParsecVote::Online(candidate)
-            | ParsecVote::PurgeCandidate(candidate)
-            | ParsecVote::RefuseCandidate(candidate)
-            | ParsecVote::RelocateResponse(candidate, _) => Some(*candidate),
-
-            ParsecVote::AddElderNode(_)
-            | ParsecVote::RemoveElderNode(_)
-            | ParsecVote::NewSectionInfo(_)
-            | ParsecVote::RelocationTrigger
-            | ParsecVote::CheckElder => None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum LocalEvent {
-    TimeoutAccept,
-    RelocationTrigger,
-    TimeoutCheckElder,
-    JoiningTimeoutResendCandidateInfo,
-    JoiningTimeoutRefused,
-    ComputeResourceProofForElder(Name, ProofSource),
-}
-
-impl LocalEvent {
-    fn to_event(&self) -> Event {
-        Event::LocalEvent(*self)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone)]
-struct InnerAction {
-    our_attributes: Attributes,
-    our_section: SectionInfo,
-    our_current_nodes: BTreeMap<Name, NodeState>,
-
-    our_votes: Vec<ParsecVote>,
-    our_rpc: Vec<Rpc>,
-    our_events: Vec<LocalEvent>,
-    our_nodes: Vec<NodeChange>,
-
-    shortest_prefix: Option<Section>,
-    node_to_relocate: Option<Node>,
-    section_members: BTreeMap<SectionInfo, Vec<Node>>,
-}
-
-impl InnerAction {
-    fn new_with_our_attributes(name: Attributes) -> Self {
-        Self {
-            our_attributes: name,
-            our_section: Default::default(),
-            our_current_nodes: Default::default(),
-
-            our_votes: Default::default(),
-            our_rpc: Default::default(),
-            our_events: Default::default(),
-            our_nodes: Default::default(),
-
-            shortest_prefix: Default::default(),
-            node_to_relocate: Default::default(),
-            section_members: Default::default(),
-        }
-    }
-
-    fn extend_current_nodes(mut self, nodes: &[NodeState]) -> Self {
-        self.our_current_nodes.extend(
-            nodes
-                .iter()
-                .map(|state| (Name(state.node.0.name), state.clone())),
-        );
-        self
-    }
-
-    fn extend_current_nodes_with(self, value: &NodeState, nodes: &[Node]) -> Self {
-        let node_states = nodes
-            .iter()
-            .map(|node| NodeState {
-                node: node.clone(),
-                ..value.clone()
-            })
-            .collect_vec();
-        self.extend_current_nodes(&node_states)
-    }
-
-    fn with_section_members(mut self, section: &SectionInfo, nodes: &[Node]) -> Self {
-        self.section_members
-            .insert(*section, nodes.iter().cloned().collect());
-        self
-    }
-
-    fn add_node(&mut self, node_state: NodeState) {
-        self.our_nodes
-            .push(NodeChange::AddResourceProofing(node_state.node));
-        self.our_current_nodes
-            .insert(Name(node_state.node.0.name), node_state);
-    }
-
-    fn remove_node(&mut self, node: Node) {
-        self.our_nodes.push(NodeChange::Remove(node));
-        self.our_current_nodes.remove(&Name(node.0.name));
-    }
-
-    fn set_relocating_state(&mut self, name: &Name) {
-        let node = &mut self.our_current_nodes.get_mut(name).unwrap();
-
-        node.is_relocating = true;
-        self.our_nodes.push(NodeChange::Relocating(node.node));
-    }
-
-    fn set_online_state(&mut self, name: &Name) {
-        let node = &mut self.our_current_nodes.get_mut(name).unwrap();
-
-        node.is_resource_proofing = false;
-        self.our_nodes.push(NodeChange::Online(node.node));
-    }
-
-    fn set_elder_state(&mut self, name: &Name, value: bool) {
-        let node = &mut self.our_current_nodes.get_mut(name).unwrap();
-
-        node.is_elder = value;
-        self.our_nodes.push(NodeChange::Elder(node.node, value));
-    }
-
-    fn set_section_info(&mut self, section: SectionInfo) {
-        self.our_section = section;
-    }
-}
-
-#[derive(Clone)]
-struct Action(Rc<RefCell<InnerAction>>);
-
-impl Action {
-    fn new(inner: InnerAction) -> Self {
-        Action(Rc::new(RefCell::new(inner)))
-    }
-
-    fn inner(&self) -> InnerAction {
-        (*self.0.borrow()).clone()
-    }
-
-    fn remove_processed_state(&self) {
-        let inner = &mut self.0.borrow_mut();
-
-        inner.our_votes.clear();
-        inner.our_rpc.clear();
-        inner.our_nodes.clear();
-        inner.our_events.clear();
-    }
-
-    fn vote_parsec(&self, vote: ParsecVote) {
-        self.0.borrow_mut().our_votes.push(vote);
-    }
-
-    fn send_rpc(&self, rpc: Rpc) {
-        self.0.borrow_mut().our_rpc.push(rpc);
-    }
-
-    fn schedule_event(&self, event: LocalEvent) {
-        self.0.borrow_mut().our_events.push(event);
-    }
-
-    fn add_node_ressource_proofing(&self, candidate: Candidate) {
-        let state = NodeState {
-            node: Node(candidate.0),
-            is_resource_proofing: true,
-            ..NodeState::default()
-        };
-        self.0.borrow_mut().add_node(state);
-    }
-
-    fn set_candidate_online_state(&self, candidate: Candidate) {
-        self.0
-            .borrow_mut()
-            .set_online_state(&Name(candidate.0.name));
-    }
-
-    fn remove_node(&self, candidate: Candidate) {
-        self.0.borrow_mut().remove_node(Node(candidate.0));
-    }
-
-    fn check_shortest_prefix(&self) -> Option<Section> {
-        self.0.borrow().shortest_prefix
-    }
-
-    fn check_elder(&self) -> Option<ChangeElder> {
-        let inner = &self.0.borrow();
-        let our_current_nodes = &inner.our_current_nodes;
-
-        let (new_elders, ex_elders, _elders) = {
-            let mut sorted_values = our_current_nodes
-                .values()
-                .cloned()
-                .sorted_by(|left, right| {
-                    left.is_relocating
-                        .cmp(&right.is_relocating)
-                        .then(left.node.0.age.cmp(&right.node.0.age).reverse())
-                        .then(left.node.0.name.cmp(&right.node.0.name))
-                })
-                .collect_vec();
-            let elder_size = std::cmp::min(3, sorted_values.len());
-            let adults = sorted_values.split_off(elder_size);
-
-            let new_elders = sorted_values
-                .iter()
-                .filter(|elder| !elder.is_elder)
-                .cloned()
-                .collect_vec();
-            let ex_elders = adults
-                .iter()
-                .filter(|elder| elder.is_elder)
-                .cloned()
-                .collect_vec();
-
-            (new_elders, ex_elders, sorted_values)
-        };
-
-        let changes = new_elders
-            .iter()
-            .map(|elder| (elder, true))
-            .chain(ex_elders.iter().map(|elder| (elder, false)))
-            .map(|(elder, new_is_elder)| (elder.node, new_is_elder))
-            .collect_vec();
-
-        if changes.is_empty() {
-            None
-        } else {
-            Some(ChangeElder {
-                changes,
-                new_section: SectionInfo(inner.our_section.0, inner.our_section.1 + 1),
-            })
-        }
-    }
-
-    fn get_elder_change_votes(&self, change_elder: &ChangeElder) -> Vec<ParsecVote> {
-        change_elder
-            .changes
-            .iter()
-            .map(|(node, new_is_elder)| match new_is_elder {
-                true => ParsecVote::AddElderNode(*node),
-                false => ParsecVote::RemoveElderNode(*node),
-            })
-            .chain(std::iter::once(ParsecVote::NewSectionInfo(
-                change_elder.new_section,
-            )))
-            .collect_vec()
-    }
-
-    fn mark_elder_change(&self, change_elder: ChangeElder) {
-        for (node, new_is_elder) in &change_elder.changes {
-            self.0
-                .borrow_mut()
-                .set_elder_state(&Name(node.0.name), *new_is_elder);
-        }
-        self.0
-            .borrow_mut()
-            .set_section_info(change_elder.new_section);
-    }
-
-    fn get_relocating_candidate(&self) -> Candidate {
-        let inner = &self.0.borrow();
-
-        if let Some(relocating) = inner
-            .our_current_nodes
-            .values()
-            .find(|state| state.is_relocating)
-        {
-            return Candidate(relocating.node.0);
-        }
-
-        match &inner.node_to_relocate {
-            Some(Node(val)) => Candidate(*val),
-            None => panic!("node_to_relocate not setup"),
-        }
-    }
-
-    fn is_elder_state(&self, candidate: Candidate) -> bool {
-        self.0
-            .borrow()
-            .our_current_nodes
-            .get(&Name(candidate.0.name))
-            .unwrap()
-            .is_elder
-    }
-
-    fn is_candidate_relocating_state(&self, candidate: Candidate) -> bool {
-        self.0
-            .borrow()
-            .our_current_nodes
-            .get(&Name(candidate.0.name))
-            .unwrap()
-            .is_relocating
-    }
-
-    fn is_our_name(&self, name: Name) -> bool {
-        self.our_name() == name
-    }
-
-    fn our_name(&self) -> Name {
-        Name(self.0.borrow().our_attributes.name)
-    }
-
-    fn send_node_approval_rpc(&self, candidate: Candidate) {
-        let section = GenesisPfxInfo(self.0.borrow().our_section);
-        self.send_rpc(Rpc::NodeApproval(candidate, section));
-    }
-
-    fn set_candidate_relocating_state(&self, candidate: Candidate) {
-        self.0
-            .borrow_mut()
-            .set_relocating_state(&Name(candidate.0.name));
-    }
-
-    fn send_relocate_response_rpc(&self, candidate: Candidate) {
-        let section = self.0.borrow().our_section;
-        self.send_rpc(Rpc::RelocateResponse(candidate, section));
-    }
-
-    fn send_candidate_relocated_info(&self, candidate: Candidate, section: SectionInfo) {
-        self.send_rpc(Rpc::RelocatedInfo(candidate, section));
-    }
-
-    fn send_candidate_proof_request(&self, candidate: Candidate) {
-        let source = self.our_name();
-        let proof = ProofRequest { value: source.0 };
-        self.send_rpc(Rpc::ResourceProof {
-            candidate,
-            proof,
-            source,
-        });
-    }
-
-    fn send_candidate_proof_receipt(&self, candidate: Candidate) {
-        let source = self.our_name();
-        self.send_rpc(Rpc::ResourceProofReceipt { candidate, source });
-    }
-
-    fn start_compute_resource_proof(&self, source: Name, proof: ProofRequest) {
-        self.schedule_event(LocalEvent::ComputeResourceProofForElder(
-            source,
-            ProofSource(2),
-        ));
-    }
-
-    fn get_section_members(&self, section_info: &SectionInfo) -> Vec<Node> {
-        self.0
-            .borrow()
-            .section_members
-            .get(section_info)
-            .unwrap()
-            .clone()
-    }
-
-    fn send_connection_info_request(&self, destination: Name) {
-        let source = self.our_name();
-        self.send_rpc(Rpc::ConnectionInfoRequest {
-            source,
-            destination,
-            connection_info: source.0,
-        });
-    }
-
-    fn send_candidate_info(&self, destination: Name) {
-        let candidate = Candidate(self.0.borrow().our_attributes);
-        self.send_rpc(Rpc::CandidateInfo {
-            candidate,
-            destination,
-            valid: true,
-        });
-    }
-
-    fn send_resource_proof_response(&self, destination: Name, proof: Proof) {
-        let candidate = Candidate(self.0.borrow().our_attributes);
-        self.send_rpc(Rpc::ResourceProofResponse {
-            candidate,
-            destination,
-            proof,
-        });
-    }
-}
-
-impl Default for Action {
-    fn default() -> Action {
-        Action::new(InnerAction::new_with_our_attributes(Attributes::default()))
-    }
-}
-
-impl Debug for Action {
-    fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
-        self.0.borrow().fmt(formatter)
-    }
-}
-
-impl PartialEq for Action {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.borrow().eq(&*other.0.borrow())
-    }
-}
 
 #[derive(Debug, PartialEq, Default, Clone)]
 struct CheckAndProcessElderChangeState {
